Check installation configuration statusesf, small improvements

+++ distributions/osx/install.py
@@ -14,8 +14,7 @@ def at_exit():
 
 def check_system_requirements():
     is_good = True
-    required_utils = [("brew", "https://treehouse.github.io/installation-guides/mac/homebrew"),
-                      ("make", "brew install make"),
+    required_utils = [("make", "brew install make"),
                       ("openssl", "brew install openssl")]
 
     for util, help_str in required_utils:
@@ -23,7 +22,11 @@ def check_system_requirements():
         if res.returncode:
             print("{} is required but not installed. Installation instruction: {}".format(util, help_str))
             is_good = False
+    # we need brew only if not all requirements are met
     if not is_good:
+        res = subprocess.run(['which', 'brew'], stdout=subprocess.PIPE)
+        if res.returncode:
+            print("brew is not installed. How to install: /usr/local/opt/openssl@1.1/bin/openssl")
         sys.exit("please install required utils and try again")
 
 check_system_requirements()
@@ -67,8 +70,12 @@ def download_file(url):
     return filename
 
 def get_openssl_path():
-    res = subprocess.check_output(["brew", "--prefix", "openssl"])
-    return res.decode('utf-8').rstrip()
+    res = subprocess.check_output(["which", "openssl"])
+    res = res.decode('utf-8').rstrip()
+    # delete /bin/openssl from path
+    # from /usr/local/opt/openssl@1.1/bin/openssl -> /usr/local/opt/openssl@1.1
+    root_dir = '/'.join(res.split('/')[:-2])
+    return root_dir
 
 try:
     make_dir(INSTALL_DIR)
@@ -94,17 +101,21 @@ print("configuring python: {}".format(config_cmd))
 print("this may take a while...")
 print("CPPFLAGS: {}".format(os.getenv("CPPFLAGS", "")))
 print("LDFLAGS: {}".format(os.getenv("LDFLAGS", "")))
-os.system(config_cmd)
+status = os.system(config_cmd)
+if status:
+    sys.exit("python configuring process finished with error(s)")
 print("installing python to %s" % PYTHON_DIR)
 print("this may take a while...")
 
-os.system("cd {}; make install --silent".format(PYTHON_SRC_DIR))
+status = os.system("cd {}; make install --silent".format(PYTHON_SRC_DIR))
+if status:
+    sys.exit("installation process finished with error(s)")
+
 shutil.rmtree(PYTHON_SRC_DIR)
 
 PYTHON_EXEC = os.path.join(PYTHON_DIR, 'bin/python3')
 print("python location: {}".format(PYTHON_EXEC))
 os.system("{} -V".format(PYTHON_EXEC))
-
 
 os.system('{} -m pip install torch==1.5.0 torchvision==0.6.0 -f https://download.pytorch.org/whl/torch_stable.html --no-warn-script-location'.format(PYTHON_EXEC))
 if VERSION == '':
